feat: Renamed component commands to library and added alias for library topic as lib
