+++ cli/src/_templates/library/new/gitignore.ejs.t
@@ -3,4 +3,3 @@ to: <%= name %>/.gitignore
 ---
 node_modules/
 dist/
-.tooljet/
