+++ cli/src/_templates/library/new/globaldts.ejs.t
