+++ cli/src/_templates/library/new/hello-world.ejs.t
@@ -5,7 +5,7 @@ import React from "react";
 import { ToolJet } from "@tooljet/custom-component-sdk";
 
 export const HelloWorld: React.FC = () => {
-  const [firstName, setFirstName] = ToolJet.useStateString({ name: 'firstName', label: 'First Name', initialValue: 'John' });
+  const [firstName] = ToolJet.useStateString({ name: 'firstName', label: 'First Name', initialValue: 'John' });
 
   return (
     <div>
