+++ cli/src/_templates/library/new/index.ejs.t
