+++ cli/src/_templates/library/new/tsconfig.ejs.t
