+++ cli/src/lib/library/auth.ts
