+++ cli/src/lib/library/manifest-generator.ts
